backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
SQLAlchemy==1.4.23
python-dotenv==0.19.0
Werkzeug==2.0.3
tencentcloud-sdk-python
requests>=2.25.1
openai>=1.0.0